Factor ClientMainloopThread out into a header too.
