+++ apps/CMakeLists.txt
@@ -26,6 +26,7 @@ target_link_libraries(osvr_calibrate jsoncpp_lib osvrClientKitCpp osvrCommon eig
 add_executable(osvr_reset_yaw
     osvr_reset_yaw.cpp
     ClientMainloop.h
+    ClientMainloopThread.h
     RecomposeTransform.h
     WrapRoute.h)
 target_link_libraries(osvr_reset_yaw jsoncpp_lib osvrClientKitCpp osvrClient osvrTransform osvrRouting osvrCommon eigen-headers boost_thread ${Boost_PROGRAM_OPTIONS_LIBRARIES})
